pyproject.toml
[tool.poetry]
name = "spark-surveymonkey"
version = "0.1.0"
description = "Convert SurveyMonkey respones in JSON to a PySpark DataFrame"
authors = ["Emanuel Ferm <emanuel@eferm.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6.1"
jq = "^1.0.1"

[tool.poetry.dev-dependencies]
py4j = "^0.10.9"
jupyter = "^1.0.0"
requests = "^2.23.0"
pytest = "^5.4.3"
python-dotenv = "^0.13.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
